# Get index of first non ascii char