@kunalspathak 2020-08-10
Do you mind sending a PR to `dotnet/performance` to add `IsNormalized_GetIndexOfFirstNonAsciiChar ()` benchmark? We need it to track .NET 3.1 vs. .NET 5 performance improvements in this area.