@pgovind 2020-08-13
Filed https://github.com/dotnet/runtime/issues/40805 to follow up on Tamar's suggestions here. Considering that we're already seeing decent perf improvements here, how do folks feel about merging this PR now and investigating the suggestions in a future PR?