@kunalspathak 2020-08-13
Couple of things:
- There was some feedback given in https://github.com/dotnet/runtime/pull/39103#discussion_r455552959 that I don't see we implemented. It was also brought up by Tamar, the key point is we should not mimic the SSE2 logic of `AddSaturate()` because ARM64 has better instructions to do similar operation. I agree that it will take some time to make it most efficient implementation, but we should at least think about ways to do the operation without `AddSaturate()`, etc. Currently, we are doing more than needed operations like 3 `AddPairwise` which might show improvement on WSL2 but might have degraded performance on different processors.
- If I see the code of `IsNormalized_GetIndexOfFirstNonAsciiChar()` in https://github.com/dotnet/performance/pull/1445/files#diff-2d837d38e3d94ab0d7f80e232693857eR90 , I see that you are testing it on ascii data instead of non-ascii data. Was that intentional? Are the above results from the ascii version as seen in the PR in performance repo?

With that, I am not sure if we should rush this in for RC1.